Added an example to demonstrate template-based source code generation.
The "mist_gen-01" example uses MiST Engine to generate different types of
documents from the same data but different templates. The data is the
information about some imaginary C functions: names, types of the
arguments, etc.

Using the 3 different sets of templates, different resulting documents can
be generated from that data:
- a C source file with the 'stubs' for these functions;
- a verbose plain text report describing the functions;
- a html page with a table of functions.

The example pretends to demonstrate the power of strict model-view
separation:

one set of data + several sets of templates = different resulting documents

If you need to change the look of the result, usually it is enough to
change only the templates to achieve the desired result.

You may experiment with this example providing a new set of templates and
generating a completely different documents from that data.

The example is written in C++ and can be built both on Linux and on
Microsoft Windows (with MinGW). GCC C++ compiler 4.0 or newer is
recommended.

This also allows to demonstrate how to use MiST engine API from a C++
application. MiST Engine is used here as a shared object / DLL, which is
one of the two recommended ways to use it (the other one is to use
'mist_engine' command line tool).

Note that the way this example operates resembles T2C to some extent but is
much simpler.

See README file in the example for other details.