An example to demonstrate generation of payload source using templates.
The example is similar to "controller_payload" but the source code of the
payload modules is generated using the templates (MiST Engine is used for
this). The collected data is output via trace events rather than printk in
"controller_payload".

Definitions of trace events are also generated using the templates.