Fixed: execute_process() was used incorrectly in CMakeLists.txt
If more than one command is specified in a single execute_process()
statement, it is as if these commands were executed in a pipeline by
the shell: command1 | command2 | ...

The commands in a pipeline MAY execute concurrently.

This commit fixes, for example, a race condition in the following construct:

execute_process (
    COMMAND rm -rf "${KEDR_GEN_INSTALL_PREFIX}"
    COMMAND mkdir -p "${KEDR_GEN_INSTALL_PREFIX}"
)

Here, 'rm' could remove the directory AFTER 'mkdir' has created it =>
strange failures of the configuration phase.

Several cosmetic changes have been applied as well.