conf: added a check for stack trace reliability
It is now checked during configuration phase, whether the system supports
reliable stack traces. To be exact, it is checked if the kernel is built
with support for frame pointers and/or stack unwind info on.

Some parts of KEDR need reliable stack traces, support for 'caller_address'
in fault simulation scenarios is one of the examples.