LeakCheck: handling of allocs and frees is now deferred via a work queue
Using a work queue allows to significantly reduce the time the replacement
functions spend with a spinlock held.

The internal API was revisited in several ways, not only to make use of
the work queue but also to hide some implementation-specific stuff that was
in the headers before (changes in our stack trace-related API made it
possible).

Various minor improvements were made as well.