Change behaviour of capture trace program concerned reading of the trace file in the blocking mode.
Now, capture trace's default behaviour is non-blocking reading of the trace. So, program will exit when it has to wait message in the trace.
Passing option '-b' turn reading of the trace to blocking. So, program will wait message in the trace if it is needed.
In that mode SIGINT signal simply turn reading mode to non-blocking.

Documentation for the capture trace will be changed later.