tests: Test for stack trace support is now enabled by default
The test for kedr_save_stack_trace() is now enabled even if the system does
not support reliable stack traces. Now it just allows to check whether the
function crashes or not, not far beyond that.

In addition, a warning is now printed during the configuration stage if the
latter is the case.