Added support for processing of add_to_page_cache_* functions
The following functions can now be intercepted and processed by KEDR:
- add_to_page_cache_locked
- add_to_page_cache_lru

add_to_page_cache* calls are handled similar to free_pages() to avoid
false positives when analysing NFS modules or the like. These modules
allocate memory pages and sometimes transfer the responsibility for
them to the kernel proper among other things.

For now, call interception tests for these functions are not provided. In
the future, they should be prepared - as soon as we figure out how to create
a reasonably simple code to use these functions in the correct environment
and how to properly clean up after that.