LeakCheck: Added definition of the new API, revisited the build system
This is the first stage of LeakCheck redesign.
The goal is to separate LeakCheck into the core (the analysis engine) and
the plugin modules that track the calls to the functions of interest and
inform the core about the allocation and deallocation events using the new
API.

The core should be capable of processing of allocation/deallocation events
from many sources simultaneously. This will help in the future when kedr is
redesigned itself to be more "service-like". As for the plugins to
LeakCheck, they should currently be capable of tracking the events in only
one target module at a time, for simplicity.

LeakCheck API can also be used by other kernel-mode components to obtain
more accurate results in some special cases. For example, a component
intercepting the callback operations registered by the target module with
the kernel could use that to indicate that a given memory allocation or
deallocation has occurred there. It is clear now that interception of the
calls to the exported functions alone is not enough in some cases to
properly track memory allocation/deallocation. LeakCheck API allows using
other components to improve accuracy of LeakCheck.

The redesign should also make it easier to write custom tools using
LeakCheck (e.g. tracking the functions not tracked by the standard plugin to
LeakCheck).

In this commit, the declarations of the API have been provided along with
the stubs for their implementations. The .data files have been updated to
use the new API.

The build system has been revisited to accomodate the new design of
LeakCheck. The LeakCheck core should be implemented later.