Fixed: The tests should use /var/tmp/kedr tree for work than the build tree
Before this commit, many tests used subdirectories of their build directory
to mount debugfs for example. On some systems, this lead to test failures
because debugfs failed to unmount / remount sometimes.

Specifically, on Fedora 17 with GNOME 3, each time debugfs was mounted
somewhere in the build tree (usually under /home/<user>/somewhere),
gnome-shell-hotplug-sniffer was launched and started to access the files
in the mounted debugfs somehow. When it reached 'trace_pipe' maintained by
Ftrace, it hung. As the file was considered used, the system did not allow
to unmount the corresponding directory.

It seems that gnome-shell-hotplug-sniffer does not process the directories
in /var/tmp tree, so it is better to create the working directories for the
tests there.

This commit adjusts the tests to use the appropriate subdirectories of
/var/tmp/kedr rather than the places in the build tree.