LeakCheck: Adjusted comparisons in the AWK script used in the tests
Some Linux distros (e.g. Debian 6) use mawk as awk utility.

mawk 1.3.3 seems to consider "0xffff88002ada0600" equal to
"0xffff88002ada0b80" if one or both are the values of the variables set
with -v and are compared with '==' as they are.

To find out whether a given variant of AWK has this problem, one can use
the following test script:

------- cut here ----------
#!/bin/awk -f
BEGIN {
    if (aaa == bbb) {
        printf("Equal: %s and %s\n", aaa, bbb)
    }
    else {
        printf("Not equal: %s and %s\n", aaa, bbb)
    }
}
-------- cut here ---------

Execute the script as follows:
    awk -f <script_file> -v aaa=0xffff88002ada0600 -v bbb=0xffff88002ada0b80

This peculiarity led to sporadic failures of the tests depending on which
memory addresses were involved there.

Comparison of these strings with "t" appended seems to produce correct
results, so I have updated the script to use that workaround.