build: Make the standard plugins for call monitoring disabled by default
This facility provided by KEDR seems to be used very rarely, so it seems
reasonable to turn it off by default to speed up build, etc.

Note that the support fo call monitoring itself is NOT disabled this way,
only the plugins that process the calls to the particular functions are.

The user-provided plugins that perform call monitoring will also work no
matter if KEDR_STANDARD_CALLM_PAYLOADS is set to "OFF" or not.