LeakCheck: Revisited handling of "unallocated free" events
There are situations, where "unallocated frees" reported by LeakCheck are
currently inevitable. One of the examples is 'readahead' callback provided
by some file system modules (btrfs, ceph, ...) where the memory is actually
allocated by the kernel proper.

If there are many such false "unallocated frees", the data LeakCheck stores
for them may occupy a significant amount of memory (about 80-160 bytes is
needed for each event).

To avoid wasting memory this way, a limit has been imposed on the number of
the groups of such events (a group consists of the events with the same call
stack). Parameter 'bad_free_groups_stored' defines that limit.

In addition, the "unallocated free" events are now combined into groups not
when the target is about to unload but rather as soon as they are detected.