LeakCheck: Relaxed comparison of call stacks
On many systems, the "outermost" call stack elements for a system call may
differ for different processes. If the kernel-space parts of two call stacks
are equal, LeakCheck will now consider the call stacks equal even if such
"system call gate" elements are different. This allows to combine the
corresponding records in LeakCheck's reports and this way prevent the bloat
of these reports.