LeakCheck: Allow flushing stats at any time
It is now possible to force LeakCheck to output the information about the
not freed memory at any time, not only when the target module is unloaded. This
may result in false positives, e.g. for the memory blocks that are only freed in
the exit function of the target, but can still be handy in some cases.

LeakCheck now maintains "flush" file in <debugfs>/<target_name>/. If the user
writes anything to that file, the results of LeakCheck will be available
immediately thereafter in the relevant files in debugfs and/or in the system
log.

The idea and the original version of the patch were proposed by Maxim Malkov.
Details and discussion: http://code.google.com/p/kedr/issues/detail?id=9