LeakCheck: Clear the collected data if requested
A file named "clear" is now provided for each target in the appropriate
directory in debugfs.

If the user writes anything to this file, LeakCheck will "forget" the
information about memory allocations and deallocations collected so far.

Can be useful if the user is not interested in the memory allocations made by
the target module before a particular moment, e.g. before the target is loaded
and fully initialized.

Likely be used in conjunction with flushing of the results.

Details: http://code.google.com/p/kedr/issues/detail?id=9